Implement MCS protocol
The MCS layer receives PDUs and sends them to a router. Routers keep
track of the connected users and route the appropriate PDUs to them.
When a user joins a new channel, a factory is called to create a
MCSChannel object that will handle communication for that (userID,
channelID) pair. This way, the next layers can just call a send function
and the appropriate user ID and channel ID will be set automatically,
which helps isolate the MCS layer. Parsing for connection settings (e.g:
GCC) is not done in the MCS layer: a callback to an observer is made
when receiving connection PDUs to delegate this job. Determining whether
or not a channel ID is valid is also left to the channel factory. This
seems to be the least complicated way to implement it.