Version restrictions on dependencies (discussed in #219)
Aimed to be able to run from distro-packaged alternatives back to 2018
(Debian, Ubuntu 18.04 LTS). Not tested so I might be off but we will
adjust if we get bug reports.

Twisted and service-identity run on CalVer and have multi-year
deprecation policies. I think it's better not to bound them too much.